fix(bridge): temperature 改用 f64，修掉精度雜訊的根因
MessagesRequest.temperature 原本是 f32。JSON 數字本來就是 f64，先窄化
成 f32 再經 json!() 加寬回去會產生 0.30000001192092896，送到上游就不再
是使用者輸入的值。改成 f64 讓它全程不轉換，同時移除翻譯層裡
to_string().parse() 的繞道寫法。

+++ src-tauri/src/bridge/anthropic/request.rs
@@ -18,7 +18,10 @@ pub struct MessagesRequest {
     #[serde(default)]
     pub stream: Option<bool>,
     #[serde(default)]
-    pub temperature: Option<f32>,
+    /// 用 f64 而非 f32：JSON 數字本來就是 f64，先窄化成 f32 再加寬回去會
+    /// 夾帶精度雜訊（`0.3f32 as f64` = 0.30000001192092896），送到上游就
+    /// 不再是使用者輸入的值。
+    pub temperature: Option<f64>,
     #[serde(default)]
     pub stop_sequences: Option<Vec<String>>,
     #[serde(default)]
